Use new minimum SDK version interface
