+++ hyperion-geiger-counter/src/main/java/com/willowtreeapps/hyperion/geigercounter/GeigerCounterPlugin.java
@@ -1,10 +1,8 @@
 package com.willowtreeapps.hyperion.geigercounter;
 
-import android.app.Activity;
 import android.content.Context;
 import android.content.res.AssetManager;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.Display;
 import android.view.WindowManager;
 
@@ -12,15 +10,23 @@ import com.google.auto.service.AutoService;
 import com.willowtreeapps.hyperion.plugin.v1.Plugin;
 import com.willowtreeapps.hyperion.plugin.v1.PluginModule;
 
+import static android.os.Build.VERSION_CODES.JELLY_BEAN;
+
 @AutoService(Plugin.class)
+@SuppressWarnings("NewApi")
 public class GeigerCounterPlugin extends Plugin {
 
+    static final int API_VERSION = JELLY_BEAN;
     static final String LOG_TAG = "Hyperion Geiger Counter";
 
-    @Nullable
     private static DroppedFrameObserver observer;
 
     // Plugin
+
+    @Override
+    protected int minimumRequiredApi() {
+        return API_VERSION;
+    }
 
     protected void onApplicationCreated(@NonNull Context context) {
         super.onApplicationCreated(context);
@@ -30,7 +36,7 @@ public class GeigerCounterPlugin extends Plugin {
         WindowManager windowManager = (WindowManager) context.getSystemService(Context.WINDOW_SERVICE);
         Display display = windowManager.getDefaultDisplay();
 
-        observer = DroppedFrameObserverFactory.getInstance(assetManager, display);
+        observer = new DroppedFrameObserver(assetManager, display);
     }
 
     @Override
